move contact.py to item.py. No need for a separate file and contacts need to be included in the stac item by the make_item anyway
